Separate SetCookie header policy
