+++ lib/protocol/http/header/cookie.rb
@@ -3,7 +3,6 @@
 # Released under the MIT License.
 # Copyright, 2019-2026, by Samuel Williams.
 
-require_relative "multiple"
 require_relative "../cookie"
 
 module Protocol
@@ -21,7 +20,7 @@ module Protocol
 				def self.parse(value)
 					self.new([value])
 				end
-
+				
 				# Coerces a value into a parsed header object.
 				#
 				# @parameter value [String | Array] the value to coerce.
@@ -34,18 +33,18 @@ module Protocol
 						self.parse(value.to_s)
 					end
 				end
-
+				
 				# Initializes the cookie header with the given values.
 				#
 				# @parameter value [Array | Nil] an array of cookie strings, or `nil` for an empty header.
 				def initialize(value = nil)
 					super()
-
+					
 					if value
 						self.concat(value)
 					end
 				end
-
+				
 				# Parses the `cookie` header into a hash of cookie names and their corresponding cookie objects.
 				#
 				# @returns [Hash(String, HTTP::Cookie)] a hash where keys are cookie names and values are {HTTP::Cookie} objects.
@@ -53,15 +52,15 @@ module Protocol
 					cookies = self.collect do |string|
 						HTTP::Cookie.parse(string)
 					end
-
+					
 					cookies.map{|cookie| [cookie.name, cookie]}.to_h
 				end
-
+				
 				# Serializes the `cookie` header by joining individual cookie strings with `"; "` per RFC 6265.
 				def to_s
 					join("; ")
 				end
-
+				
 				# Whether this header is acceptable in HTTP trailers.
 				# Cookie headers should not appear in trailers as they contain state information needed early in processing.
 				# @returns [Boolean] `false`, as cookie headers are needed during initial request processing.
@@ -69,30 +68,9 @@ module Protocol
 					false
 				end
 			end
-
-			# The `set-cookie` header sends cookies from the server to the user agent.
-			#
-			# Each `Set-Cookie` header must be a separate header field — they cannot be combined.
-			# It is used to store cookies on the client side, which are then sent back to the server
-			# in subsequent requests using the `cookie` header.
-			class SetCookie < Multiple
-				# Parses the `set-cookie` headers into a hash of cookie names and their corresponding cookie objects.
-				#
-				# @returns [Hash(String, HTTP::Cookie)] a hash where keys are cookie names and values are {HTTP::Cookie} objects.
-				def to_h
-					cookies = self.collect do |string|
-						HTTP::Cookie.parse(string)
-					end
-
-					cookies.map{|cookie| [cookie.name, cookie]}.to_h
-				end
-
-				# Whether this header is acceptable in HTTP trailers.
-				# @returns [Boolean] `false`, as set-cookie headers are needed during initial response processing.
-				def self.trailer?
-					false
-				end
-			end
 		end
 	end
 end
+
+# Preserve the existing behavior where requiring this file also defines SetCookie.
+require_relative "set_cookie"
